requirements.txt
langchain>=0.3.0
langchain-ollama>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
boto3>=1.35.0
pyyaml>=6.0
requests>=2.31.0
